use named vars instead of auto vars
b889ac0 attempted to use automatic variables whenever possible, but the
changes are flawed. flatpak-install requires REMOTE to start with either
/ or ./ for it to consider it as a name of local repository. Otherwise,
it's treated as a remote reference, which is invalid most of the time.
$< in makefiles also retain only the file name of the first prerequisite.
Due to misunderstandings and possible environment misconfiguraions,
those changes were committed.

This commit replaces all uses of automatic variable with their named
counterparts.

Fix: #1